Fix Start Browsing Button on back
The "Start Browsing Your OneDrive" button was checking for the existance
of the service object and then if it saw an expection was creating the
service object and sending the user to the root of their OneDrive.  This
unforunately would not work in the success case where a service object
existed, adding in the proper navigateToRoot call in the success flow.
Fixes #68.